ts billing-timeframe and cleanup
